RecRefacto download
Use requests-cache
Refacto yaml
Rename "field" argument in yaml to "territory"
Handle zip
Handle nested zip/7zip
Handle CSV/DBF pattern (not only shapefiles)
Refacto tests with CachedSession patching
Split download on multiple files (download, scraper, dataset)
