+++ cartiflette/download/__init__.py
@@ -4,23 +4,18 @@ from cartiflette.download.dev import (
     get_cog_year,
 )
 
+
 from cartiflette.download.download import (
-    Dataset,
-    BaseScraper,
-    HttpScraper,
-    FtpScraper,
-    MasterScraper,
     download_sources,
+    upload_vectorfile_to_s3,
+    download_all,
 )
 
 __all__ = [
     "get_vectorfile_communes_arrondissement",
     #     "get_BV",
     "get_cog_year",
-    "Dataset",
-    "BaseScraper",
-    "HttpScraper",
-    "FtpScraper",
-    "MasterScraper",
     "download_sources",
+    "upload_vectorfile_to_s3",
+    "download_all",
 ]
