+++ pyproject.toml
@@ -24,6 +24,7 @@ python-dotenv = "^1.0.0"
 python-magic-bin = "^0.4.14"
 requests-cache = "^1.1.0"
 unidecode = "^1.3.6"
+charset-normalizer = "^3.2.0"
 
 [tool.poetry.group.dev.dependencies]
 spyder = "^5.4.2"
